requirements.txt
discord.py==0.16.12
praw==6.3.1
